+++ Scripts/IndicatorCreateQueries/PanwIndicatorCreateQueries.py
